# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Examify Swagger",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/api/user/get-my-results": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves a paginated list of exam results for the logged-in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ielts-exam"
                ],
                "summary": "USER",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's exam results",
                        "schema": {
                            "$ref": "#/definitions/pb.GetExamByUserIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/user/profile": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the profile information of the currently authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "User Profile Response",
                        "schema": {
                            "$ref": "#/definitions/pb.User"
                        }
                    },
                    "409": {
                        "description": "Conflict Error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/user/update-information": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Allows a user to update their name and surname",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user's name and surname",
                "parameters": [
                    {
                        "description": "Name and Surname Update Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateUserNameSurnameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict Error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "pb.GetExamAbsResult": {
            "type": "object",
            "properties": {
                "bookName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "examId": {
                    "type": "string"
                },
                "listening": {
                    "type": "string"
                },
                "overall": {
                    "type": "string"
                },
                "reading": {
                    "type": "string"
                },
                "remainSection": {
                    "type": "string"
                },
                "remainTimeForEndExam": {
                    "type": "integer"
                },
                "speaking": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "writing": {
                    "type": "string"
                }
            }
        },
        "pb.GetExamByUserIdResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.GetExamAbsResult"
                    }
                },
                "totalPageCount": {
                    "type": "integer"
                }
            }
        },
...
        "pb.UpdateUserNameSurnameRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "pb.User": {
            "type": "object",
            "properties": {
                "chatId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "todayExamCount": {
                    "type": "integer"
                }
            }
        },
...
        "utils.AbsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}